Make /sysl a nested module (#7)
Testing has revealed that a simple dependency on options.proto may cause all of the transitive dependencies of protoc-gen-sysl to be pulled in. This is problematic, since somebody trying to annotate a protobuf is not in a position to reason about the implications of these dependencies.

One workaround is to make the dependency on options.proto weak. This prevents the import of the Go module containing it in the protoc-generated code, and hence avoids the transitive imports. However weak dependencies are discouraged and can lead to other issues like other tools missing references.

This PR proposes a cleaner solution: isolate options.proto in its own Go module with no dependencies, then import that into protoc-gen-sysl. Other repos that just want the annotations can depend only on the isolated module at runtime, and pull in the heavier protoc-gen-sysl only at build time.

Also:
- Update all tests to have a `go_package` with at least one `/` in it, as per new requirements.
- Update the Sysl dependency while we're at it.
- Regenerate options.pb.go.